feat: Implement secure prescription viewing and admin order management

+++ src/main/java/com/healthcare/repository/OrderItemRepository.java
@@ -3,4 +3,5 @@ package com.healthcare.repository;
 import com.healthcare.model.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> { }
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> { // CHANGED to Long
+}
