+++ src/main/java/com/healthcare/repository/OrderRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> { // CHANGED to Long
     List<Order> findByUser(User user);
 }
