+++ src/main/java/com/healthcare/repository/PrescriptionRepository.java
@@ -5,6 +5,6 @@ import com.healthcare.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface PrescriptionRepository extends JpaRepository<Prescription, Integer> {
+public interface PrescriptionRepository extends JpaRepository<Prescription, Long> { // CHANGED to Long
     List<Prescription> findByUser(User user);
 }
